refactor(backbone): rename CSP references and introduce CSPDarknet implementation
- Updated CSP references in BackboneManager and ConfigBackboneManager to use "csp_darknet" for consistency.
- Introduced a new CSPDarknet class with comprehensive architecture and configuration settings.
- Adjusted imports in the RTMO model files to reflect the new CSPDarknet structure.
- Updated test configurations to include the new CSPDarknet model.

These changes enhance clarity and maintainability of the backbone architecture, ensuring a more coherent integration of the CSPDarknet model.
