+++ tests/test_backbone.py
@@ -12,8 +12,8 @@ BACKBONE_CONFIGS = {
     "swin": [{"model_type": "swin", "use_pretrained": False}],
     "mobilenet_v2": [{"model_type": "mobilenet_v2", "use_pretrained": False}],
     "convnextv2": [{"model_type": "convnextv2", "use_pretrained": False}],
-    "c2f_darknet": [
-        {"model_type": "c2f_darknet", "use_pretrained": False, "size": size} for size in ["n", "s", "m", "l", "x"]
+    "csp_darknet": [
+        {"model_type": "csp_darknet", "use_pretrained": False, "size": size} for size in ["small", "medium", "large"]
     ],
 }
